小医仙助手对话功能调试成功

+++ src/main/java/me/icemoon/java/ai/langchain4j/assistant/IXiaoYiXianAgent.java
@@ -19,5 +19,5 @@ import dev.langchain4j.service.spring.AiServiceWiringMode;
 public interface IXiaoYiXianAgent {
 
     @SystemMessage(fromResource = "xiaoyixian-prompt-template.txt")
-    String chat(@MemoryId int memoryId, @UserMessage String userMessage);
+    String chat(@MemoryId Long memoryId, @UserMessage String userMessage);
 }
